@@ -3,13 +3,13 @@
   "version": "0.1.0",
   "dependencies": {
     "axios": "^0.19.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "3.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3"
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
